// src/providers/enums/matching-queue-readiness.enum.ts
export enum MatchingQueueReadiness {
  READY = 'READY',
  FUNDING_INCOMPLETE = 'FUNDING_INCOMPLETE',
  INCOMPLETE_SCHEDULING = 'INCOMPLETE_SCHEDULING',
  INCOMPLETE_VISIT_ADDRESS = 'INCOMPLETE_VISIT_ADDRESS',
  ACTIVE_OFFER = 'ACTIVE_OFFER',
  ACCEPTED_AWAITING_CONFIRMATION = 'ACCEPTED_AWAITING_CONFIRMATION',
  UNFULFILLABLE = 'UNFULFILLABLE',
  ALREADY_ASSIGNED = 'ALREADY_ASSIGNED',
}